updated article sharing hooks
