--- src/hooks/sharing/useGetSharedArticles.js
@@ -1,45 +1,0 @@
-import { useState } from "react";
-import axiosInstance from "../axios/axiosInstance";
-import { useDispatch } from "react-redux";
-import { setArticles, setArticleCount } from "../../features/sharingSlice";
-import { getItemFromLocalStorage } from "../utils/globalFunctions";
-
-export const useGetSharedArticles = () => {
-  const dispatch = useDispatch();
-  const [isPending, setIsPending] = useState(false);
-
-  const getSharedArticles = async (params) => {
-    setIsPending(true);
-
-    const token = getItemFromLocalStorage("token");
-
-    try {
-      const res = await axiosInstance.get("/shared", {
-        params,
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-type": "application/json; charset=UTF-8",
-        },
-      });
-
-      dispatch(setArticles(res.data.articles));
-      dispatch(setArticleCount(res.data.articleCount));
-
-      return res ? { ok: true } : { error: "Unable to fetch Articles" };
-    } catch (err) {
-      let error = "";
-      if (err.response) {
-        error = err?.response?.data?.message || "An error occurred.";
-      } else if (err.request) {
-        error = "Network error. Please try again later.";
-      } else {
-        error = "An error occurred. Please try again later.";
-      }
-      return { error };
-    } finally {
-      setIsPending(false);
-    }
-  };
-
-  return { getSharedArticles, isPending };
-};
